refactor: split verification.py into package (closes #180)
- Create src/valence/core/verification/ package
- constants.py: ReputationConstants
- enums.py: All verification-related enums
- models.py: ExternalSource, BeliefReference, Observation, DerivationProof
- evidence.py: Evidence class and create_evidence helper
- results.py: ResultDetails, Stake
- verification.py: Verification, Dispute, ReputationScore, and service classes
- __init__.py: Re-exports all public API for backward compatibility

Original file: 1837 lines
New structure: 6 modules with clear separation of concerns

All 120 verification tests pass.
