+++ src/valence/core/verification/verification.py
@@ -14,508 +14,35 @@ Spec references:
 
 from __future__ import annotations
 
-import hashlib
 import json
 import logging
 import math
-from dataclasses import asdict, dataclass, field
+from dataclasses import dataclass, field
 from datetime import datetime, timedelta
-from enum import StrEnum
 from typing import Any
 from uuid import UUID, uuid4
 
-from .confidence import DimensionalConfidence
-from .exceptions import NotFoundError, ValidationException
+from ..exceptions import NotFoundError, ValidationException
+from .constants import ReputationConstants
+from .enums import (
+    DisputeOutcome,
+    DisputeStatus,
+    DisputeType,
+    EvidenceContribution,
+    ResolutionMethod,
+    StakeType,
+    VerificationResult,
+    VerificationStatus,
+)
+from .evidence import Evidence
+from .results import ResultDetails, Stake
 
 logger = logging.getLogger(__name__)
 
 
 # ============================================================================
-# Constants (from REPUTATION.md)
-# ============================================================================
-
-
-class ReputationConstants:
-    """Constants for reputation calculations."""
-
-    # Base rewards
-    CONFIRMATION_BASE = 0.001
-    CONTRADICTION_BASE = 0.005
-    UNCERTAINTY_BASE = 0.0002
-
-    # Bounties
-    BOUNTY_MULTIPLIER = 0.5
-    FIRST_FINDER_BONUS = 2.0
-
-    # Penalties
-    CONTRADICTION_PENALTY_BASE = 0.003
-    FRIVOLOUS_DISPUTE_PENALTY = 0.2
-    COLLUSION_PENALTY = 0.5
-
-    # Calibration
-    CALIBRATION_BONUS_BASE = 0.01
-    CALIBRATION_PENALTY_THRESHOLD = 0.4
-
-    # Decay
-    MONTHLY_INACTIVITY_DECAY = 0.02
-    VERIFICATION_HALF_LIFE_MONTHS = 12
-
-    # Limits
-    MAX_DAILY_GAIN = 0.02
-    MAX_WEEKLY_GAIN = 0.08
-    REPUTATION_FLOOR = 0.1
-    MAX_STAKE_RATIO = 0.2
-
-    # Recovery
-    MAX_MONTHLY_RECOVERY = 0.03
-    PROBATION_DURATION_DAYS = 90
-
-    # Timing
-    VALIDATION_WINDOW_HOURS = 1
-    ACCEPTANCE_DELAY_HOURS = 24
-    DISPUTE_WINDOW_DAYS = 7
-    RESOLUTION_TIMEOUT_DAYS = 14
-    STAKE_LOCKUP_DAYS = 7
-    BOUNTY_STAKE_LOCKUP_DAYS = 14
-
-    # Rate limits
-    MAX_VERIFICATIONS_PER_DAY = 50
-    MAX_PENDING_VERIFICATIONS = 10
-    MAX_VERIFICATIONS_PER_BELIEF = 100
-    MAX_VERIFICATION_VELOCITY_PER_HOUR = 5
-
-
-# ============================================================================
-# Enums
-# ============================================================================
-
-
-class VerificationResult(StrEnum):
-    """Result of a verification."""
-
-    CONFIRMED = "confirmed"  # Evidence supports the belief
-    CONTRADICTED = "contradicted"  # Evidence refutes the belief
-    UNCERTAIN = "uncertain"  # Insufficient evidence either way
-    PARTIAL = "partial"  # Partially correct, with qualifications
-
-
-class VerificationStatus(StrEnum):
-    """Status of a verification."""
-
-    PENDING = "pending"  # Awaiting validation
-    ACCEPTED = "accepted"  # Validated and accepted
-    DISPUTED = "disputed"  # Under dispute
-    OVERTURNED = "overturned"  # Dispute overturned the verification
-    REJECTED = "rejected"  # Failed validation
-    EXPIRED = "expired"  # Timed out during validation
-
-
-class StakeType(StrEnum):
-    """Type of stake."""
-
-    STANDARD = "standard"  # Normal verification stake
-    BOUNTY = "bounty"  # Claiming a discrepancy bounty
-    CHALLENGE = "challenge"  # Challenging existing verification
-
-
-class EvidenceType(StrEnum):
-    """Type of evidence provided."""
-
-    BELIEF = "belief"  # Reference to another Valence belief
-    EXTERNAL = "external"  # External source (URL, paper, etc.)
-    OBSERVATION = "observation"  # Verifier's direct observation
-    DERIVATION = "derivation"  # Logical/mathematical proof
-    TESTIMONY = "testimony"  # Statement from another agent
-
-
-class EvidenceContribution(StrEnum):
-    """How evidence contributes to verification."""
-
-    SUPPORTS = "supports"  # Evidence for the belief
-    CONTRADICTS = "contradicts"  # Evidence against the belief
-    CONTEXT = "context"  # Adds relevant context
-    QUALIFIES = "qualifies"  # Adds conditions/limitations
-
-
-class ContradictionType(StrEnum):
-    """Type of contradiction found."""
-
-    FACTUALLY_FALSE = "factually_false"
-    OUTDATED = "outdated"
-    MISATTRIBUTED = "misattributed"
-    OVERSTATED = "overstated"
-    MISSING_CONTEXT = "missing_context"
-    LOGICAL_ERROR = "logical_error"
-
-
-class UncertaintyReason(StrEnum):
-    """Reason for uncertain result."""
-
-    INSUFFICIENT_EVIDENCE = "insufficient_evidence"
-    CONFLICTING_SOURCES = "conflicting_sources"
-    OUTSIDE_EXPERTISE = "outside_expertise"
-    UNFALSIFIABLE = "unfalsifiable"
-    REQUIRES_EXPERIMENT = "requires_experiment"
-
-
-class DisputeType(StrEnum):
-    """Type of dispute."""
-
-    EVIDENCE_INVALID = "evidence_invalid"
-    EVIDENCE_FABRICATED = "evidence_fabricated"
-    EVIDENCE_INSUFFICIENT = "evidence_insufficient"
-    REASONING_FLAWED = "reasoning_flawed"
-    CONFLICT_OF_INTEREST = "conflict_of_interest"
-    NEW_EVIDENCE = "new_evidence"
-
-
-class DisputeOutcome(StrEnum):
-    """Outcome of a dispute resolution."""
-
-    UPHELD = "upheld"  # Original verification stands
-    OVERTURNED = "overturned"  # Verification was wrong
-    MODIFIED = "modified"  # Result changed
-    DISMISSED = "dismissed"  # Dispute was frivolous
-
-
-class DisputeStatus(StrEnum):
-    """Status of a dispute."""
-
-    PENDING = "pending"  # Awaiting resolution
-    RESOLVED = "resolved"  # Resolution complete
-    EXPIRED = "expired"  # Timed out
-
-
-class ResolutionMethod(StrEnum):
-    """Method used to resolve a dispute."""
-
-    AUTOMATIC = "automatic"  # Algorithm decides
-    JURY = "jury"  # Random selection of jurors
-    EXPERT = "expert"  # Domain experts decide
-    APPEAL = "appeal"  # Higher-level review
-
-
-# ============================================================================
 # Data Models
 # ============================================================================
-
-
-@dataclass
-class ExternalSource:
-    """External evidence source."""
-
-    url: str | None = None
-    doi: str | None = None
-    isbn: str | None = None
-    citation: str | None = None
-    archive_hash: str | None = None
-    archived_at: datetime | None = None
-    source_reputation: float = 0.5
-
-    def to_dict(self) -> dict[str, Any]:
-        result = {k: v for k, v in asdict(self).items() if v is not None}
-        if self.archived_at:
-            result["archived_at"] = self.archived_at.isoformat()
-        return result
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> ExternalSource:
-        if data.get("archived_at") and isinstance(data["archived_at"], str):
-            data["archived_at"] = datetime.fromisoformat(data["archived_at"])
-        return cls(**data)
-
-
-@dataclass
-class BeliefReference:
-    """Reference to another belief as evidence."""
-
-    belief_id: UUID
-    holder_id: str  # DID
-    content_hash: str
-    confidence_at_time: DimensionalConfidence | None = None
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "belief_id": str(self.belief_id),
-            "holder_id": self.holder_id,
-            "content_hash": self.content_hash,
-            "confidence_at_time": (self.confidence_at_time.to_dict() if self.confidence_at_time else None),
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> BeliefReference:
-        confidence = None
-        if data.get("confidence_at_time"):
-            confidence = DimensionalConfidence.from_dict(data["confidence_at_time"])
-        return cls(
-            belief_id=(UUID(data["belief_id"]) if isinstance(data["belief_id"], str) else data["belief_id"]),
-            holder_id=data["holder_id"],
-            content_hash=data["content_hash"],
-            confidence_at_time=confidence,
-        )
-
-
-@dataclass
-class Observation:
-    """Direct observation evidence."""
-
-    description: str
-    timestamp: datetime
-    method: str
-    reproducible: bool = False
-    reproduction_instructions: str | None = None
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "description": self.description,
-            "timestamp": self.timestamp.isoformat(),
-            "method": self.method,
-            "reproducible": self.reproducible,
-            "reproduction_instructions": self.reproduction_instructions,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> Observation:
-        timestamp = data["timestamp"]
-        if isinstance(timestamp, str):
-            timestamp = datetime.fromisoformat(timestamp)
-        return cls(
-            description=data["description"],
-            timestamp=timestamp,
-            method=data["method"],
-            reproducible=data.get("reproducible", False),
-            reproduction_instructions=data.get("reproduction_instructions"),
-        )
-
-
-@dataclass
-class DerivationProof:
-    """Logical derivation as evidence."""
-
-    premises: list[UUID]
-    logic_type: str  # 'deductive', 'inductive', 'abductive'
-    proof_steps: list[str]
-    formal_notation: str | None = None
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "premises": [str(p) for p in self.premises],
-            "logic_type": self.logic_type,
-            "proof_steps": self.proof_steps,
-            "formal_notation": self.formal_notation,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> DerivationProof:
-        return cls(
-            premises=[UUID(p) if isinstance(p, str) else p for p in data["premises"]],
-            logic_type=data["logic_type"],
-            proof_steps=data["proof_steps"],
-            formal_notation=data.get("formal_notation"),
-        )
-
-
-@dataclass
-class Evidence:
-    """Evidence supporting a verification."""
-
-    id: UUID
-    type: EvidenceType
-    relevance: float  # 0.0-1.0
-    contribution: EvidenceContribution
-    verifier_notes: str | None = None
-
-    # Content (exactly one populated based on type)
-    belief_reference: BeliefReference | None = None
-    external_source: ExternalSource | None = None
-    observation: Observation | None = None
-    derivation: DerivationProof | None = None
-    testimony_statement: str | None = None
-
-    def __post_init__(self):
-        if self.relevance < 0.0 or self.relevance > 1.0:
-            raise ValidationException(
-                "Evidence relevance must be between 0.0 and 1.0",
-                "relevance",
-                self.relevance,
-            )
-
-    def content_hash(self) -> str:
-        """Compute SHA-256 hash of evidence content."""
-        content = json.dumps(self.to_dict(), sort_keys=True)
-        return hashlib.sha256(content.encode()).hexdigest()
-
-    def to_dict(self) -> dict[str, Any]:
-        result: dict[str, Any] = {
-            "id": str(self.id),
-            "type": self.type.value,
-            "relevance": self.relevance,
-            "contribution": self.contribution.value,
-            "verifier_notes": self.verifier_notes,
-        }
-        if self.belief_reference:
-            result["belief_reference"] = self.belief_reference.to_dict()
-        if self.external_source:
-            result["external_source"] = self.external_source.to_dict()
-        if self.observation:
-            result["observation"] = self.observation.to_dict()
-        if self.derivation:
-            result["derivation"] = self.derivation.to_dict()
-        if self.testimony_statement:
-            result["testimony_statement"] = self.testimony_statement
-        return result
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> Evidence:
-        evidence_id = data.get("id")
-        if evidence_id and isinstance(evidence_id, str):
-            evidence_id = UUID(evidence_id)
-        elif not evidence_id:
-            evidence_id = uuid4()
-
-        belief_ref = None
-        if data.get("belief_reference"):
-            belief_ref = BeliefReference.from_dict(data["belief_reference"])
-
-        external = None
-        if data.get("external_source"):
-            external = ExternalSource.from_dict(data["external_source"])
-
-        obs = None
-        if data.get("observation"):
-            obs = Observation.from_dict(data["observation"])
-
-        deriv = None
-        if data.get("derivation"):
-            deriv = DerivationProof.from_dict(data["derivation"])
-
-        return cls(
-            id=evidence_id,
-            type=EvidenceType(data["type"]),
-            relevance=float(data["relevance"]),
-            contribution=EvidenceContribution(data["contribution"]),
-            verifier_notes=data.get("verifier_notes"),
-            belief_reference=belief_ref,
-            external_source=external,
-            observation=obs,
-            derivation=deriv,
-            testimony_statement=data.get("testimony_statement"),
-        )
-
-
-@dataclass
-class ResultDetails:
-    """Structured breakdown of verification result."""
-
-    # For CONFIRMED
-    confirmation_strength: str | None = None  # 'strong', 'moderate', 'weak'
-    confirmed_aspects: list[str] = field(default_factory=list)
-
-    # For CONTRADICTED
-    contradiction_type: ContradictionType | None = None
-    corrected_belief: str | None = None
-    severity: str | None = None  # 'minor', 'moderate', 'major', 'critical'
-
-    # For PARTIAL
-    accurate_portions: list[str] = field(default_factory=list)
-    inaccurate_portions: list[str] = field(default_factory=list)
-    accuracy_estimate: float | None = None  # 0.0-1.0
-
-    # For UNCERTAIN
-    uncertainty_reason: UncertaintyReason | None = None
-    additional_evidence_needed: list[str] = field(default_factory=list)
-
-    # Universal
-    confidence_assessment: list[dict[str, Any]] = field(default_factory=list)
-
-    def to_dict(self) -> dict[str, Any]:
-        result: dict[str, Any] = {}
-        if self.confirmation_strength:
-            result["confirmation_strength"] = self.confirmation_strength
-        if self.confirmed_aspects:
-            result["confirmed_aspects"] = self.confirmed_aspects
-        if self.contradiction_type:
-            result["contradiction_type"] = self.contradiction_type.value
-        if self.corrected_belief:
-            result["corrected_belief"] = self.corrected_belief
-        if self.severity:
-            result["severity"] = self.severity
-        if self.accurate_portions:
-            result["accurate_portions"] = self.accurate_portions
-        if self.inaccurate_portions:
-            result["inaccurate_portions"] = self.inaccurate_portions
-        if self.accuracy_estimate is not None:
-            result["accuracy_estimate"] = self.accuracy_estimate
-        if self.uncertainty_reason:
-            result["uncertainty_reason"] = self.uncertainty_reason.value
-        if self.additional_evidence_needed:
-            result["additional_evidence_needed"] = self.additional_evidence_needed
-        if self.confidence_assessment:
-            result["confidence_assessment"] = self.confidence_assessment
-        return result
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> ResultDetails:
-        contradiction_type = None
-        if data.get("contradiction_type"):
-            contradiction_type = ContradictionType(data["contradiction_type"])
-
-        uncertainty_reason = None
-        if data.get("uncertainty_reason"):
-            uncertainty_reason = UncertaintyReason(data["uncertainty_reason"])
-
-        return cls(
-            confirmation_strength=data.get("confirmation_strength"),
-            confirmed_aspects=data.get("confirmed_aspects", []),
-            contradiction_type=contradiction_type,
-            corrected_belief=data.get("corrected_belief"),
-            severity=data.get("severity"),
-            accurate_portions=data.get("accurate_portions", []),
-            inaccurate_portions=data.get("inaccurate_portions", []),
-            accuracy_estimate=data.get("accuracy_estimate"),
-            uncertainty_reason=uncertainty_reason,
-            additional_evidence_needed=data.get("additional_evidence_needed", []),
-            confidence_assessment=data.get("confidence_assessment", []),
-        )
-
-
-@dataclass
-class Stake:
-    """Reputation stake for a verification."""
-
-    amount: float
-    type: StakeType
-    locked_until: datetime
-    escrow_id: UUID
-
-    def __post_init__(self):
-        if self.amount < 0:
-            raise ValidationException("Stake amount cannot be negative", "amount", self.amount)
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "amount": self.amount,
-            "type": self.type.value,
-            "locked_until": self.locked_until.isoformat(),
-            "escrow_id": str(self.escrow_id),
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> Stake:
-        locked_until = data["locked_until"]
-        if isinstance(locked_until, str):
-            locked_until = datetime.fromisoformat(locked_until)
-
-        escrow_id = data["escrow_id"]
-        if isinstance(escrow_id, str):
-            escrow_id = UUID(escrow_id)
-
-        return cls(
-            amount=float(data["amount"]),
-            type=StakeType(data["type"]),
-            locked_until=locked_until,
-            escrow_id=escrow_id,
-        )
 
 
 @dataclass
@@ -1188,7 +715,7 @@ def validate_evidence_requirements(
 
     # Check for invalid evidence
     for e in evidence:
-        if e.type == EvidenceType.BELIEF and e.belief_reference:
+        if e.type.name == "BELIEF" and e.belief_reference:
             # Circular reference check would need belief_id from context
             pass
 
@@ -1237,7 +764,7 @@ def validate_dispute_submission(
 
 
 # ============================================================================
-# Verification Service Functions
+# Verification Service
 # ============================================================================
 
 
@@ -1792,46 +1319,3 @@ class VerificationService:
         logger.info(f"Reputation update for {reputation.identity_id}: {old_value:.4f} -> {new_value:.4f} ({delta:+.4f}) - {reason}")
 
         return event
-
-
-# ============================================================================
-# Module-level convenience functions
-# ============================================================================
-
-
-def create_evidence(
-    evidence_type: EvidenceType,
-    contribution: EvidenceContribution,
-    relevance: float = 0.8,
-    **kwargs: Any,
-) -> Evidence:
-    """Convenience function to create evidence.
-
-    Args:
-        evidence_type: Type of evidence
-        contribution: How it contributes
-        relevance: Relevance score
-        **kwargs: Type-specific arguments
-
-    Returns:
-        Evidence object
-    """
-    evidence = Evidence(
-        id=uuid4(),
-        type=evidence_type,
-        relevance=relevance,
-        contribution=contribution,
-    )
-
-    if evidence_type == EvidenceType.EXTERNAL:
-        evidence.external_source = ExternalSource(**kwargs)
-    elif evidence_type == EvidenceType.BELIEF:
-        evidence.belief_reference = BeliefReference(**kwargs)
-    elif evidence_type == EvidenceType.OBSERVATION:
-        evidence.observation = Observation(**kwargs)
-    elif evidence_type == EvidenceType.DERIVATION:
-        evidence.derivation = DerivationProof(**kwargs)
-    elif evidence_type == EvidenceType.TESTIMONY:
-        evidence.testimony_statement = kwargs.get("statement", "")
-
-    return evidence
